i'm kind of an idiot (chore: move icons logic to build time)
