+++ src/components/PortGrid.svelte
@@ -56,7 +56,6 @@
       <slot name="no-results"></slot>
     {:else if portGrid && portGrid.length > 0}
       {#each portGrid as fields (fields.key)}
-        <!-- <div style={`display: ${portGrid.includes(fields) ? "grid" : "none"}`}> -->
         <PortCard
           title={fields.name}
           link={fields["is-userstyle"]
@@ -65,7 +64,6 @@
           icon={fields.icon}
           categories={fields.categories}
           color={fields.color} />
-        <!-- </div> -->
       {/each}
     {/if}
   {/key}
